[GENERATED] Initial Commit
This commit is the output of the following command (`rails@6.0.2.2`):

```
rails new select-your-own-seat \
  --webpack=stimulus \
  --database=postgresql \
  --skip-action-cable \
  --skip-action-mailbox \
  --skip-action-mailer \
  --skip-action-text \
  --skip-active-storage
```

This command skips several Rails-provided dependencies that the project
won't be utilizing.

It sets [PostgreSQL][] as the application's database, configures the
project's JavaScript to be built with [Webpack][] through [Webpacker][],
and makes [StimulusJS][] available on the project's client-side.

[PostgreSQL]: https://www.postgresql.org/
[Webpack]: https://webpack.js.org/
[Webpacker]: https://github.com/rails/webpacker
[StimulusJS]: https://stimulusjs.org/