Install JavaScript dependencies during setup
In order to execute the `webpacker:install` command executed during the
project's `rails new` command execution, the development environment
required a semi-current version of [NodeJS][] available on the system.

The [ASDF Version Manager][asdf-vm] is a useful tool that manages
versions of system-level dependencies, like Ruby and [NodeJS][].

To integrate with [ASDF][asdf-vm], this commit declares a
[`.tool-versions` file][.tool-versions] that specifies the [NodeJS][]
version to be `11.7.0`.

The `rails new` command that generated this project also generated a
`yarn install` line in the project's [`bin/setup`][bin/setup] script
that is initially commented out.

To ensure that JavaScript dependencies are available to the development
environment, this commit uncomments that line in the
[`bin/setup`][bin/setup] script to also run [`yarn
install`][yarn-install] when installing the application's dependencies.

[NodeJS]: https://nodejs.dev/
[asdf-vm]: https://asdf-vm.com/#/core-manage-asdf-vm
[.tool-versions]: #.tool-versions
[bin/setup]: #bin/setup
[yarn-install]: https://yarnpkg.com/cli/install