Add `seats#show` route and action
Render the contents of the `seats#show` controller action into the same
template as `seats#index`, but embedded the template within a block
wrapped by a browser-native [`<dialog>`][dialog] element.

To achieve this, pass [`template: "seats/index"`][template], along with
[the template's `local_assigns`][local_assigns] into a call to `render`
from within
[`app/views/seats/show.html.erb`][app/views/seats/show.html.erb].

Next, to render the `<dialog>` into the right portion of the page, nest
the elements as the block passed to a
[`content_for(:dialog)`][content_for] invocation.

A corresponding `yield(:dialog)` is declared at an approrpriate
spot in the [Application
layout][app/views/layouts/application.html.erb].

![seat show page](https://user-images.githubusercontent.com/2575027/77251202-7624cd80-6c23-11ea-8227-e4f483a8cb67.png)

The modal `<dialog>` element
---

When rendering the `seats#show` template, specify that the wrapping
`<dialog>` element has the [`open`][open] attribute.

When initially rendered (and in environments with JavaScript disabled),
the `<dialog>` element will be overlaid on top of the rest of the page.

In JavaScript-capable environments, within [`turbolinks:load`
events][turbolinks:load], find `<dialog>` elements and ensure their
browser compatibility by wrapping them in calls to the
[`dialog-polyfill`-provided `registerDialog`][dialog-polyfill].

Then, for elements that specify the `open` attribute, re-set them to be
closed, then manually invoke [`dialog.showModal()`][showModal] to render
the dialog modally.

![opening and closing the modal](https://user-images.githubusercontent.com/2575027/77251203-76bd6400-6c23-11ea-844e-7e3868214404.gif)

[template]: https://guides.rubyonrails.org/v6.0/layouts_and_rendering.html#rendering-an-action-s-template-from-another-controller
[local_assigns]: https://api.rubyonrails.org/v6.0/classes/ActionView/Template.html#method-i-local_assigns
[nested-layouts]: https://guides.rubyonrails.org/v6.0/layouts_and_rendering.html#using-nested-layouts
[content_for]: https://api.rubyonrails.org/v6.0/classes/ActionView/Helpers/CaptureHelper.html#method-i-content_for
[app/views/layouts/application.html.erb]: #app/views/layouts/application.html.erb
[dialog]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/dialog
[dialog-polyfill]: https://github.com/GoogleChrome/dialog-polyfill
[open]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/dialog#Attributes
[showModal]: https://developer.mozilla.org/en-US/docs/Web/API/HTMLDialogElement/showModal
[turbolinks:load]: https://github.com/turbolinks/turbolinks/tree/v5.2.0#full-list-of-events
[app/views/seats/show.html.erb]: #app/views/seats/show.html.erb