Unselect a Seat to remove it from visitor's Cart
Change the cart summary's `<button>` element to be rendered with a
[`button_to`][button-to] helper, transforming its underlying element to
`<form>` with an `method` attribute set to `DELETE`, and the `action`
attribute set to `/seats/:seat_id/selections/:id`.

This commit also introduces the `SelectionsController#destroy` action to
handle requests made from the `<form>` generated by the `button_to`
call.

[button-to]: https://api.rubyonrails.org/v5.2/classes/ActionView/Helpers/UrlHelper.html#method-i-button_to