Unselect a Seat from the `<dialog>`
When viewing the details for a `Seat` that is already included in the
visitor's `Cart`, present them with a button to remove it.

To determine whether or not the visitor has added a seat, check if the
Seat record’s id is in the list of the Cart record’s selected seats
(accessed via the [`has_many :seats-generated` seats_id
method][has_many]).

`DELETE` request routing
---

In order to support the removal of a Seat without looking up the
intermediary `SeatSelection` record, utilize the previously-declared
[singular `resource :seat_selection` route
declaration][singular-resource].

As previously stated, instead of routing `DELETE
/seats/:seat_id/selections/:id` requests to the `selections#destroy`
controller action, route `DELETE /seats/:seat_id/selections` requests.

In the controller action, the correct `SeatSelection` record can be
determined based on the context provided by the `Current.cart` and the
`params[:seat_id]`.

Desktop
---

![unselect from dialog](https://user-images.githubusercontent.com/2575027/77386366-03ba0780-6d61-11ea-961b-7f198dd90f82.gif)

[has_many]: https://api.rubyonrails.org/v6.0/classes/ActiveRecord/Associations/ClassMethods.html#method-i-has_many
[singular-resource]: https://guides.rubyonrails.org/v6.0/routing.html#singular-resources