Unselect a Seat from the `<dialog>`
When viewing the details for a `Seat` that is already included in the
visitor's `Cart`, present them with a button to remove it.

In order to support the removal of a Seat without looking up the
intermediary `SeatSelection` record, modify the `resources
:seat_selections` route declaration's `destroy` action to be a singular
resource instead.

This means that instead of handling `DELETE
/seats/:seat_id/selections/:id` requests, the `selections#destroy`
controller action will handle `DELETE /seats/:seat_id/selections`
requests, determining the correct `SeatSelection` record based on the
context from the `Current.cart` and the `params[:seat_id]`.

[singular-resource]: https://guides.rubyonrails.org/v5.2/routing.html#singular-resources