Switch floors from map
Add a list of `<a>` elements based on a venue's `Floor` records.

Introduce the `active_link_to` gem so that `<a>` elements that link to
the current page are styled differently.

Additionally, now that Seats can be selected from different floors,
embed the Seat's Section name in the Cart once selected.