Switch floors from map
Add a list of [`<a>` elements][mdn-a] based on a Venue record's
associated `Floor` records.

Introduce the [`active_link_to` gem][active_link_to] so that `<a>`
elements that link to the current page are styled differently.

Additionally, now that Seats can be selected from different floors,
embed the `Seat` record's `Section` name in the `Cart` once selected.

Desktop
---

![floor-switch](https://user-images.githubusercontent.com/2575027/77386337-f8ff7280-6d60-11ea-8833-db1fa4c33b68.gif)

[mdn-a]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/a
[active_link_to]: https://github.com/comfy/active_link_to/tree/v1.0.5#super-simple-example