Filter Seats based on their Price
This commit adds static HTML to render a set of `<input type="radio">`
buttons styled with `<label>` elements that wrap `<svg>` elements.

The values of those filters are static for the time being, and the
selection of a filter does not yet affect the page. The values have been
chosen as `500` cents ($5.00), `1000` cents ($10.00), and `1500` cents
($15.00). They're underscored to resemble their price in dollars, and
are iterated over with [`Array#each_cons`][each_cons] to iterate over
pairs consecutively.

Where there is a `nil` value for `next_maximum`, that indicates that
it's the highest possible value, so the `<input>` should have the
[`checked` attribute][checked].

[each_cons]: https://ruby-doc.org/core-2.6.3/Enumerable.html#method-i-each_cons
[checked]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/Input/radio#checked