Handle Price Filtering `<form>` reset events
When the "Clear Filters" button is clicked, the [`<form>`
element][mdn-form] will return to its initial state, and a
[`reset`][reset] event will be fired.

This commit configures the `<form>` element to route that event to the
`price-filter#apply` action.

Unfortunately, this event's `currentTarget` property does not refer to
the [`<input type="radio">` button element][mdn-radio] that the value
was reset to.

To account for this, consider the case when `currentTarget.value` is
`null` or `undefined` as a proxy for the `<form>` being reset.

When the element annotated with `[data-price-filter-maximum]` is set to
the empty String, none of its descendant `Section` elements will match
the CSS rule, so no visual filters will apply.

Desktop
---

![price filtering form reset](https://user-images.githubusercontent.com/2575027/77386347-fb61cc80-6d60-11ea-99ac-1adb3bae07ba.gif)

[mdn-form]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/form
[reset]: https://developer.mozilla.org/en-US/docs/Web/API/HTMLFormElement/reset_event
[mdn-radio]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input/radio
[infinity]: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Infinity