Handle Price Filtering `<form>` reset events
When the "Clear Filters" button is clicked, the `<form>` element will
return to its initial state, and a [`reset`][reset] event will be fired.

This commit configures the `<form>` element to route that event to the
`seats#filterSeats` action.

Unfortunately, this event's `currentTarget`
property does not refer to the `<input type="radio">` button that the
value was reset to.

To account for this, handle the case when `currentTarget.value` is
`undefined` as a proxy for the `<form>` being reset, and replace it with
a value that will _never_ be less than a price on the page (e.g.
[`Infinity`][infinity]).

[reset]: https://developer.mozilla.org/en-US/docs/Web/API/HTMLFormElement/reset_event
[infinity]: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Infinity