Add Zoom control overlay
This commit introduces the HTML that will control the map's zoom level.
The controls add `<button>` elements that contain inline-`<svg>`
plus and minus icons.

The elements are declared with [`type="button"`][button]. While this
attribute feels redundant, it's important to declare, since we expect to
extend them with JavaScript behavior.

According to the MDN Documentation, `type="button"` ensures that they
won't submit `<form>` elements, should they find themselves
in a situation where `<form>` elements are their ancestors:

> `type="submit"`: The button submits the form data to the server. This
> is the default if the attribute is not specified, or if the attribute
> is dynamically changed to an empty or invalid value.

> `type="button"`: The button has no default behavior. It can have
> client-side scripts associated with the element's events, which are
> triggered when the events occur.

Hiding elements that require JavaScript
---

Since the `svgPanZoom` package will be responsible for zooming in and
out of the map, that behavior will only be available in
JavaScript-capable browser environments.

To hide the zoom controls when JavaScript is disabled, declare their
parent element with the [`hidden` attribute][hidden].

When JavaScript _is enabled_, retrieve the zoom controls element
(annotated with the [`[data-target="seats.zoomControls"]`
attribute][stimulus-target]), remove the [`hidden` attribute][hidden
when the `seats` controller attaches to the DOM in its
[`connect()`][connect] lifecycle callback.

[hidden]: https://developer.mozilla.org/en-US/docs/Web/HTML/Global_attributes/hidden
[stimulus-target]: https://stimulusjs.org/reference/targets
[connect]: https://stimulusjs.org/reference/lifecycle-callbacks

Desktop
---
![map zoom buttons](https://user-images.githubusercontent.com/2575027/77386362-03217100-6d61-11ea-92b0-b902be52b175.png)

[button]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/button#attr-type