Pan and zoom within the SVG map element
In the spirit of improving the user experience, this commit enables
panning and zooming around the seat map. To do so, this commit adds a
dependency on the [`svg-pan-zoom` package][svg-pan-zoom].

To manage the configuration and execution of that plugin, this commit
introduces `controllers/seats_controller` JavaScript module. That module
will be instantiated with [elements that declare
`data-controller="seats"`][controllers].

Once an element with `data-controller="seats"` exists on the page, the
module's [`connect()` life cycle hook][connect] will be invoked, which
will [initialize the `svgPanZoom` plugin
configuration][svg-pan-zoom-usage].

The `svgPanZoom` function expects either a [CSS selector][css-selector]
referencing the [`<svg>` element][mdn-svg] to make pan-zoom capable, or
a reference to the element itself. To retrieve that element, the `<svg>`
element that the page treats as a Map is [annotated with the
`data-target="seats.map"` attribute][stimulus-target], referenced by
calling `this.mapTarget`, and passed as an argument to the `svgPanZoom`
function.

When the element is removed from the page, the module invokes the
[`disconnect()` life cycle hook][disconnect]. "Disconnection" is the
perfect time to [call `.destroy()` on the `this.map` instance][destroy],
which deconstructs the `svgPanZoom` instance.

System Test driver
---

Since this commit is the first in a series of commits that will be
[progressively enhancing][mdn-progressive] behavior when in
JavaScript-capable browsing environments, re-declare the
`ApplicationSystemTestCase` class' [`driven_by :selenium`
declaration][driven_by] so that the test harness exercises the
application code in a realistic test environment.

Desktop
---

![panning-zooming-click](https://user-images.githubusercontent.com/2575027/77386342-fa309f80-6d60-11ea-84b5-8b1e49386809.gif)

[controllers]: https://stimulusjs.org/reference/controllers
[connect]: https://stimulusjs.org/reference/lifecycle-callbacks#connection
[svg-pan-zoom]: https://github.com/ariutta/svg-pan-zoom/tree/3.6.0
[svg-pan-zoom-usage]: https://github.com/ariutta/svg-pan-zoom/tree/3.6.1#how-to-use
[css-selector]: https://developer.mozilla.org/en-US/docs/Web/CSS/CSS_Selectors
[mdn-svg]: https://developer.mozilla.org/en-US/docs/Web/SVG/Element/svg
[stimulus-target]: https://stimulusjs.org/reference/targets
[disconnect]: https://stimulusjs.org/reference/lifecycle-callbacks#disconnection
[destroy]: https://github.com/ariutta/svg-pan-zoom/tree/3.6.1#public-api
[mdn-progressive]: https://developer.mozilla.org/en-US/docs/Glossary/Progressive_Enhancement
[driven_by]: https://api.rubyonrails.org/v6.0/classes/ActionDispatch/SystemTestCase.html#method-c-driven_by