Assess what we've achieved without JavaScript
Up until this point in this example application's development
life-cycle, we've been able to achieve an end-to-end Seat selection
experience, including:

* browsing the available Seats, within a Section, on a Floor, inside a
  Venue.

* viewing informative details about an available Seat in a modal dialog.

* dismissing that modal dialog and browsing other Seats' information.

* adding a Seat to the Cart from within that modal dialog.

* indicating which Seats are currently selected, and which Seats are
  still available.

* removing a selected Seat from the Cart.

* filtering the available Seats, emphasizing those within the User's
  price range.

The entirety of this behavior has been implemented in a way that is
fully functional in a browsing environment that does not have JavaScript
enabled.

Execute System Tests without JavaScript
---

To prove that point, this commit demonstrates that's it's possible to
execute the application's [System Test suite][system-tests] in a
JavaScript-less environment.

This change is purely for the sake of demonstrating parity between the
JavaScript-capable browsing experience and a JavaScript-disabled
experience.

In a production application, the System Test suite could be a mixture of
Selenium-driven and Rack-driven tests. Striking a good balance between
the two is an exercise left up to the teams managing those applications.

[Progressive Enhancement][]
---

The underlying philosophy enabling this style of web application
implementation is called [Progressive Enhancement][]:

> **Progressive enhancement** is a design philosophy that centers around
> providing a baseline of essential content and functionality to as many
> users as possible, while at the same time going further and delivering
> the best possible experience only to users of the most modern browsers
> that can run all the required code.

One side-effect of  developing the application with an implicit
JavaScript-less constraint is that our implementation has favored
simplicity over interactivity. That isn't to say that the current user
experience is unpleasant (or does not provide value), rather that there
are plenty of additional steps we could take to push the limits of the
application much, much further.

Where do we go from here?
===

The application is functional, and provides enough value to users that
it could serve as a feature in a <dfn
id="minimum-viable-product">Minimum Viable Product (<abbr title="Minimum
Viable Product">MVP</abbr>)</dfn>.

That isn't to say that it's perfect. There are various facets of the
user experience, application performance, and the underlying
implementation that could be improved by the presence of
JavaScript-powered interactivity.

Subsequent commits will improve and extend existing functionality,
[progressively enhancing][Progressive Enhancement] the foundation that
we've built thus far.

When reading the rest of the commits, please consider what it would be
like for each one to serve as the final commit in a series of
enhancements.

Since source code is a liability and not an asset, evaluate the
trade-offs between the additional code in the diff weighed against
perceived user value and the potential cost of maintaining that code
into the future.

[system-tests]: https://guides.rubyonrails.org/v6.0/testing.html#system-testing
[rack_test]: https://github.com/teamcapybara/capybara/tree/3.21.0#racktest
[Progressive Enhancement]: https://developer.mozilla.org/en-US/docs/Glossary/Progressive_Enhancement