Introduce a Loading Indicator
This commit adds the HTML necessary to display a loading indicator over
the map and various other controls on the page.

The indicator contains an inline-rendered `<svg>` element that we
animate to spin indefinitely.

Future commits will add logic to dynamically hide and show the indicator
by adding and removing the `.is-loading` CSS class.