Introduce a Loading Indicator
This commit adds the HTML necessary to display a loading indicator over
the map and various other controls on the page.

The indicator contains an inline-rendered [`<svg>` element][mdn-svg]
that we animate to spin indefinitely.

Future commits will add logic to dynamically hide and show the indicator
by adding and removing the `.is-loading` CSS class.

Desktop
---

![loading-spinner](https://user-images.githubusercontent.com/2575027/77386339-f9980900-6d60-11ea-995a-4ce0e25f5ec0.gif)

[mdn-svg]: https://developer.mozilla.org/en-US/docs/Web/SVG/Element/svg