Dynamically Hide and Show Loading Spinner
Add the `loadingOverlay` target to the `seats` controller, and [route
the `document`-wide global events][@document] named
`turbolinks:request-start` and `turbolinks:load` to the
`seats#loadingStarted` and `seats#loadingFinished` actions respectively.

When handling [`turbolinks:request-start` and `turbolinks:load`
events][turbolinks-events], only hide and show the loading indicator
when we're discarding our map element, since it's persisted across
navigations when the `data-turbolinks-permanent` attribute is present.

The CSS class that visually indicates a loading state is `.is-loading`.
This class is passed as the `[data-seats-loading-class]` attribute on
the `[data-controller="seats"]` element in the [`seats/index`
template][app/views/seats/index.html.erb]. Once declared as such, that
value is available through Stimulus' [Data Maps
interface][stimulus-data].

Desktop
---

![loading-spinner-on-nav](https://user-images.githubusercontent.com/2575027/77386338-f8ff7280-6d60-11ea-909c-365cfaa3a2d5.gif)

[@document]: https://stimulusjs.org/reference/actions#global-events
[turbolinks-events]: https://github.com/turbolinks/turbolinks/tree/v5.2.0#full-list-of-events
[app/views/seats/index.html.erb]: #app/views/seats/index.html.erb
[stimulus-data]: https://stimulusjs.org/reference/data-maps