Hide elements that require JavaScript
This commit hides the elements that do not function without JavaScript
enabled: the zoom controls.

To hide the zoom controls when JavaScript is disabled, declare their
`<div>` with the [`hidden`][hidden] attribute.

Next, when JavaScript _is enabled_, remove the `hidden` attribute when
the `seats` controller attaches to the DOM in its [`connect()`][connect]
lifecycle callback.

[hidden]: https://developer.mozilla.org/en-US/docs/Web/HTML/Global_attributes/hidden
[connect]: https://stimulusjs.org/reference/lifecycle-callbacks