Extract the `seats/frame` partial
This commit prepares our templates to be shared across `SeatsController`
actions.

This commit replaces the `seats#index` template with the HTML rendered
by the `seats/frame` partial.

The invocation is called with `partial:` as the key that refers to
`seats/frame`. Subsequent calls to re-use the template will be made with
`layout:`, and will be invoked with a block argument.