Execute System Tests without JavaScript
Configure the `test/system` test harness to drive files.

Rack-driven tests are significantly faster to set up and execute, and
more closely mimic the end-user experience for a JavaScript-disabled
runtime.

In a production application, the System Test suite could be a mixture of
Selenium-driven and Rack-driven tests. Striking a good balance between
the two is an exercise left up to the teams managing the application.

[rack_test]: https://github.com/teamcapybara/capybara/tree/3.21.0#racktest 9e5b97699be0f9dbdc92d85a2809dfcf5523aa61